requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain==0.3.27
langchain-openai==0.3.33
nltk
inflect

